Moved content of material to medium
