+++ include/medium/properties.hpp
